Improve module architecture and add documentation
