+++ fetching/__init__.py
@@ -1,3 +1,3 @@
-from .project import repository_factory, Repository
+from .repo import repository_factory, Repository
 
 __all__ = [repository_factory.__name__, Repository.__name__]
